Create Product (name Combine modelId & colorId)

+++ src/main/java/com/menghour/java/school/phoneshopnight/repository/ModelRepository.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.menghour.java.school.phoneshopnight.entity.Model;
 
-public interface ModelRepository extends JpaRepository<Model,Integer> {
-    List<Model> findByBrandId(Integer brandId);	
-	List<Model> findByBrandIdAndName(Integer brandId, String name);
+public interface ModelRepository extends JpaRepository<Model,Long> {
+    List<Model> findByBrandId(Long brandId);	
+	List<Model> findByBrandIdAndName(Long brandId, String name);
 
 }
