+++ src/main/java/com/menghour/java/school/phoneshopnight/service/BrandService.java
@@ -7,9 +7,9 @@ import org.springframework.data.domain.Page;
 import com.menghour.java.school.phoneshopnight.entity.Brand;
 public interface BrandService {
 	Brand create(Brand brand);
-	Brand getById(Integer id);
-	Brand update(Integer id, Brand brandUpdate);
+	Brand getById(Long id);
+	Brand update(Long id, Brand brandUpdate);
 	List<Brand> getBrands(String name);
-	//List<Brand> getBrands(Map<String, String> params);
+	//List<Brand> getBrands(Map<String, String> params);	
 	Page<Brand> getBrands(Map<String, String> params);
 }
