+++ src/main/java/com/menghour/java/school/phoneshopnight/service/impl/BrandServiceImpl.java
@@ -26,12 +26,12 @@ public class BrandServiceImpl implements BrandService {
 		return brandRepository.save(brand);
 	}
 	@Override
-	public Brand getById(Integer id) {
+	public Brand getById(Long id) {
 		return brandRepository.findById(id)
 				.orElseThrow(()->new ResourceNotFoundException("Brand", id));
 	}
 	@Override
-	public Brand update(Integer id, Brand brandUpdate) {
+	public Brand update(Long id, Brand brandUpdate) {
 		Brand brand = getById(id);
 		brand.setName(brandUpdate.getName());
 		return brandRepository.save(brand);
